test/aom_random: fix int sanitizer warnings
implicit conversion from type 'int' of value 1008 (32-bit, signed)
to type 'uint8_t' (aka 'unsigned char') changed the value to 240 (8-bit, unsigned)

BUG=aomedia:2357

+++ test/acm_random.h
@@ -67,7 +67,8 @@ class ACMRandom {
     // Returns a random value near 0 or near 255, to better exercise
     // saturation behavior.
     const uint8_t r = Rand8();
-    return r < 128 ? r << 4 : r >> 4;
+    return r < 128 ? static_cast<uint8_t>(r << 4)
+                   : static_cast<uint8_t>(r >> 4);
   }
 
   int PseudoUniform(int range) { return random_.Generate(range); }
